+++ NEWS.md
@@ -1,5 +1,6 @@
 # dendroTools 0.0.7.
 
+* For regression problems with two independent variables, there are transfer functions available now. Transfer functions are given as 3D surface plots based on the plotly R package. 
 * There is a new vignette created for the compare_methods() description
 * For the optimization phase, you can now choose between RMSE and RSquared metrics for final model selection
 * All tuning parameters have new arguments, which allows users to specify a vector of possible values to be tested in the optimization process.